+++ src/kill-yank/index.ts
@@ -1,3 +1,4 @@
+import { Minibuffer } from "src/minibuffer";
 import * as vscode from "vscode";
 import { Position, Range, TextEditor } from "vscode";
 import { EditorIdentity } from "../editorIdentity";
@@ -12,6 +13,7 @@ export { AppendDirection };
 export class KillYanker implements vscode.Disposable {
   private textEditor: TextEditor;
   private killRing: KillRing | null; // If null, killRing is disabled and only clipboard is used.
+  private minibuffer: Minibuffer;
 
   private isAppending = false;
   private prevKillPositions: Position[];
@@ -23,9 +25,10 @@ export class KillYanker implements vscode.Disposable {
 
   private disposables: vscode.Disposable[];
 
-  constructor(textEditor: TextEditor, killRing: KillRing | null) {
+  constructor(textEditor: TextEditor, killRing: KillRing | null, minibuffer: Minibuffer) {
     this.textEditor = textEditor;
     this.killRing = killRing;
+    this.minibuffer = minibuffer;
 
     this.docChangedAfterYank = false;
     this.prevKillPositions = [];
@@ -114,6 +117,11 @@ export class KillYanker implements vscode.Disposable {
     const selections = this.textEditor.selections;
 
     const flattenedText = killRingEntity.asString();
+
+    if (this.minibuffer.isReading) {
+      this.minibuffer.paste(flattenedText);
+      return;
+    }
 
     if (killRingEntity.type === "editor") {
       const regionTexts = killRingEntity.getRegionTextsList();